% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaF_AUC.R
\name{deltaF_AUC}
\alias{deltaF_AUC}
\title{A function that calculates the area under the curve for a specified range in the data.}
\usage{
deltaF_AUC(dataframe, begin, finish)
}
\arguments{
\item{dataframe}{A dataframe containing intensity data of a time series on a long format - gathered under the Cell ID#}

\item{begin}{The value in the data to begin integrating from}

\item{finish}{The final value for the area you wish to integrate}
}
\description{
Uses the auc function from the MESS package to calculate the area under the curve for all the regions in a specified dataframe.
It outputs this information as a wide dataframe.
}
\examples{
auc <- integrateCurve(cellNorbegint, 50, 250)

}
\keyword{Calcium}
\keyword{DRC}
\keyword{MESS}
