% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeltafDivFoNormalizeData.R
\name{DeltafDivFoNormalizeData}
\alias{DeltafDivFoNormalizeData}
\title{A function to normalize calcium-sensitivie dye data}
\usage{
DeltafDivFoNormalizeData(file)
}
\arguments{
\item{file}{Path to a csv file that contains the time-series data for multiple cells obtained in ImageJ}
}
\description{
A function to normalize calcium-sensitivie dye data
}
\examples{
CalciumExperiment191 <- minmaxCalciumData('./Data/Wholeimage_Calcium_Experiment191.csv')

Example of storing data:
wholeNormInt191 <- wholeCalciumExperiment191[[1]]
longWholeNormInt191 <-wholeCalciumExperiment191[[2]]
}
\keyword{Calcium}
\keyword{Imaging}
