% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdeltafoNormalizeData.R
\name{fdeltafoNormalizeData}
\alias{fdeltafoNormalizeData}
\title{A function to normalize calcium-sensitivie dye data}
\usage{
fdeltafoNormalizeData(file)
}
\arguments{
\item{file}{Path to a csv file that contains the time-series data for multiple cells obtained in ImageJ}
}
\description{
This is a replacement to the deprecated 'normalizeCalciumData' function.
It is still included in the package because I don't want to have to redo any old code that may use that function.
}
\details{
This was created because it is more explicit about the method of normalization and I anticipate having other normalization functions in the future.

This function will take the data obtained from ImageJ time-series and normalize them to each ROI.
It cleans up the data as well by removing the time column and rawIntDen column obtained from the ImageJ csv.
It also renames the columns from IntDenx to Cell x.

It sets the maximum intensity in the time series to 1, and the minimum to 0.
This allows the ROI's analyzed to be compared to one another.

The output of the function should be stored as a variable.
There are two outputs for this function:
One is a standard wide-format that is easy to look at manually.
The other is a 'long' format where the intensity data has been gathered under a key for the cell/ROI being analyzed.

Normalization method: (x-min(x)) / (max(x)-min(x)
}
\examples{
CalciumExperiment191 <- minmaxCalciumData('./Data/Wholeimage_Calcium_Experiment191.csv')

Example of storing data:
wholeNormInt191 <- wholeCalciumExperiment191[[1]]
longWholeNormInt191 <-wholeCalciumExperiment191[[2]]
}
\keyword{Calcium}
\keyword{Imaging}
