% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupsData.R
\name{groupsData}
\alias{groupsData}
\title{A function that generates a tidy dataframe that groups the Cells/ROIs based upon the strength of their response at a specific drug concentration.}
\usage{
groupsData(dataframe, firstquartile, thirdquartile)
}
\arguments{
\item{dataframe}{A tidy dataframe that has}

\item{firstquartile}{Value of the first quartile for whatever concentration you want to separate responses by}

\item{thirdquartile}{Value of the third quartile for whatever concentration you want to separate responses by}
}
\description{
This way when visualizing the individual data points on the DRC, you can see if a strong early response leads to a strong late respopnse
}
\examples{
groupsData(longcellnormint, 0.073, 0.32)

}
\keyword{Calcium}
\keyword{DRC}
