% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importRawData.R
\name{importRawData}
\alias{importRawData}
\title{A function to import the raw calcium imaging data}
\usage{
importRawData(file)
}
\arguments{
\item{file}{Path to a csv file that contains the time-series data for multiple cells obtained in ImageJ}
}
\description{
This function will take the data obtained from ImageJ time-series without normalization.
It cleans up the data as well by removing the area and IntDen1 columns.
It also renames the columns from RawIntDenx to Cell x.
}
\details{
The output of the function should be stored as a variable.
There are two outputs for this function:
One is a standard wide-format that is easy to look at manually.
The other is a 'long' format where the intensity data has been gathered under a key for the cell/ROI being analyzed.
}
\examples{
CalciumExperiment191 <- importRawData('./Data/Wholeimage_Calcium_Experiment191.csv')

Example of storing data:
wholeNormInt191 <- wholeCalciumExperiment191[[1]]
longWholeNormInt191 <-wholeCalciumExperiment191[[2]]
}
\keyword{Calcium}
\keyword{Imaging}
