% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importRawIntDen.R
\name{importRawIntDen}
\alias{importRawIntDen}
\title{A function to calculate mean fluorescence intensity from}
\usage{
importRawIntDen(file)
}
\arguments{
\item{file}{Path to a csv file that contains the time-series data for multiple cells obtained in ImageJ}
}
\description{
This function will take the data obtained from ImageJ time-series and normalize them to each ROI.
It cleans up the data as well by removing the time column and rawIntDen column obtained from the ImageJ csv.
It also renames the columns from IntDenx to Cell x.
}
\details{
It sets the maximum intensity in the time series to 1, and the minimum to 0.
This allows the ROI's analyzed to be compared to one another.

The output of the function should be stored as a variable.
There are two outputs for this function:
One is a standard wide-format that is easy to look at manually.
The other is a 'long' format where the intensity data has been gathered under a key for the cell/ROI being analyzed.

Normalization method: (x-min(x)) / (max(x)-min(x)
}
\examples{
CalciumExperiment191 <- normalizeCalciumData('./Data/Wholeimage_Calcium_Experiment191.csv')

Example of storing data:
wholetempRawCalcium191 <- wholeCalciumExperiment191[[1]]
longWholetempRawCalcium191 <-wholeCalciumExperiment191[[2]]
}
\keyword{Calcium}
\keyword{Imaging}
