% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDRC.R
\name{plotDRC}
\alias{plotDRC}
\title{A function that generates a dose-response curve for cellbased calcium data (or any multi-ROI based data).}
\usage{
plotDRC(dataframe, maxdose)
}
\arguments{
\item{dataframe}{A tidy dataframe that has}

\item{maxdose}{The highest dose used in the dose response curve}
}
\description{
A function that generates a dose-response curve for cellbased calcium data (or any multi-ROI based data).
}
\examples{
plotDRC(longcellnormint.grp, 10)

}
\keyword{Calcium}
\keyword{DRC}
\keyword{MESS}
