% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsCellCalciumData.R
\name{plotsCellCalciumData}
\alias{plotsCellCalciumData}
\title{A function to plate normalized calcium data}
\usage{
plotsCellCalciumData(data, conditions = c(0), columns)
}
\arguments{
\item{data}{A dataframe containing intensity data of a time series on a long format - gathered under the Cell ID#}

\item{conditions}{A simple list or vector of timepoints where the conditions the cells were exposed to were changed.}

\item{columns}{An integer representing the number of columns you would like your plot to have.}
}
\description{
This function will plot the calcium data for individual ROI's.
It can also be used to plot the whole field of view (if data is available for that).
But it was designed to be able to simultaneously plot all the individual cell-based ROI's.
}
\examples{
conditions191 <- c(50,100,150,230)
plotsCellCalciumData(longWholeNormInt191, conditions191, 5)
}
\keyword{calcium}
\keyword{ggplot2}
\keyword{plot}
