% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateF_Fo.R
\name{integrateF_Fo}
\alias{integrateF_Fo}
\title{A function used to quantify the response in the region of interest normalized by F/Fo.}
\usage{
integrateF_Fo(data, conditions, averageWindow)
}
\arguments{
\item{data}{= Dataframe containing fluorescence intensity data}

\item{conditions}{= vectory of integers indicating when}

\item{averageWindow}{= the window to use prior to stimulus for calculating Fo

i = odd index of conditions (corresponds of start of new condition)
j = index of response}
}
\description{
It normalizes the intensity data during stimulus window by calculating an average Fo before stimulus, and then dviding the fluorescence values in this window by the background Fo.
}
\details{
Normalizing first by an averaged background (Fo) right before stimulus corrects for photobleaching/drift.



Inputs:
}
