% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantDeltaF_Fo.R
\name{quantDeltaF_Fo}
\alias{quantDeltaF_Fo}
\title{A function used to quantify the response in the region of interest normalized by F/Fo.}
\usage{
quantDeltaF_Fo(data, conditions, averageWindow)
}
\description{
It normalizes the intensity data during stimulus window by calculating an average Fo before stimulus, and then dviding the fluorescence values in this window by the background Fo.
}
\details{
Normalizing first by an averaged background (Fo) right before stimulus corrects for photobleaching/drift.



Inputs:

i = odd index of conditions (corresponds of start of new condition)
j = index of response
}
